Fixed the unit tests.
Change-Id: I6da0bb00147667975e72f38bf7c0e04044105582

Uncommented the get external REST data since the unit test needed it to
ensure a connection with the init source data. The DB was failing
because the quick and dirty test checked that an event_id of 1 returned
but was throwing an error when the DB was empty. I changed it to insert
a record, then get the recored by safetyreportid, check to ensure that
the get recored wasn't null, and the roll-back the insertion. 

Change-Id: I6da0bb00147667975e72f38bf7c0e04044105582